+++ src/user-lists/user-list.service.ts
@@ -12,6 +12,8 @@ import { PageMetaDto } from "../common/dtos/page-meta.dto";
 import { HighlightOptionsDto } from "../highlight/dtos/highlight-options.dto";
 import { DbUserHighlight } from "../user/entities/user-highlight.entity";
 import { GetPrevDateISOString } from "../common/util/datetimes";
+import { WorkspaceService } from "../workspace/workspace.service";
+import { DbWorkspaceUserLists } from "../workspace/entities/workspace-user-list.entity";
 import { CreateUserListDto } from "./dtos/create-user-list.dto";
 import { DbUserList } from "./entities/user-list.entity";
 import { DbUserListContributor } from "./entities/user-list-contributor.entity";
@@ -29,7 +31,10 @@ export class UserListService {
     private userHighlightRepository: Repository<DbUserHighlight>,
     @InjectRepository(DbUser, "ApiConnection")
     private userRepository: Repository<DbUser>,
-    private pagerService: PagerService
+    @InjectRepository(DbWorkspaceUserLists, "ApiConnection")
+    private workspaceUserListsRepository: Repository<DbWorkspaceUserLists>,
+    private pagerService: PagerService,
+    private workspaceService: WorkspaceService
   ) {}
 
   baseQueryBuilder(): SelectQueryBuilder<DbUserList> {
@@ -114,14 +119,27 @@ export class UserListService {
     });
   }
 
-  async addUserList(userId: number, list: CreateUserListDto) {
-    const newUserList = this.userListRepository.create({
+  async addUserList(userId: number, list: CreateUserListDto, workspaceId: string) {
+    let existingWorkspace;
+
+    if (workspaceId === "") {
+      existingWorkspace = await this.workspaceService.findPersonalWorkspaceByUserId(userId);
+    } else {
+      existingWorkspace = await this.workspaceService.findOneById(workspaceId);
+    }
+
+    const newUserList = await this.userListRepository.save({
       user_id: userId,
       name: list.name,
       is_public: list.is_public,
     });
 
-    return this.userListRepository.save(newUserList);
+    await this.workspaceUserListsRepository.save({
+      user_list_id: newUserList.id,
+      workspace_id: existingWorkspace.id,
+    });
+
+    return newUserList;
   }
 
   async addUserListContributor(listId: string, userId: number, username?: string) {
@@ -179,6 +197,18 @@ export class UserListService {
   }
 
   async deleteUserList(listId: string) {
+    const workspaceUserList = await this.workspaceUserListsRepository.findOne({
+      where: {
+        user_list_id: listId,
+      },
+      withDeleted: false,
+    });
+
+    if (!workspaceUserList) {
+      throw new NotFoundException("could not find workspace user list link for given insight");
+    }
+
+    await this.workspaceUserListsRepository.softDelete(workspaceUserList.id);
     return this.userListRepository.softDelete(listId);
   }